requirements.txt
streamlit
pillow
pyheif
piexif
geopy